fuse-bridge: preserve original mode and umask
FUSE used to interpret the umask itself. That was a bad idea, since
there are cases where umask is not applied, such as when extended POSIX
ACLs are present and default ACLs are set on parent directory.

Thus, FUSE was changed to pass umask to operation handler itself in case
the fuse server is initialized with FUSE_DONT_MASK. In case FUSE
supports it, gluster client sends mode unmasked and umask separately (
masked mode is still sent, for compatibility with old servers).
the mode as-is for compatibility with older clients.

Signed-off-by: Lubomir Rintel <lubo.rintel@gooddata.com>
Tested-by: Lubomir Rintel <lubo.rintel@gooddata.com>
BUG: 3776