posix: apply original umask in case client sent it
FUSE used to interpret the umask itself. That was a bad idea, since
there are cases where umask is not applied, such as when extended POSIX
ACLs are present and default ACLs are set on parent directory.

The FUSE bridge was changed to send original mode with umask (alongside
masked mode, for compatibility). If that is the case, we set umask and
use the original mode for node creation calls (mkdir, mknod, create) and
let the underlying POSIX filesystem do the right thing, otherwise we
just follow the old behavior and apply the client-masked mode as-is
(with zero umask).

The original, broken, behavior is preserved in following cases:
* Unpatched client (not sending umask with original mode)
* Unpatched server (not understanding umask with original mode)
* Old FUSE on client side (FUSE < 7.12 or linux < 2.6.31)
  (can not find out the umask and original mode)

Signed-off-by: Lubomir Rintel <lubo.rintel@gooddata.com>
Tested-by: Lubomir Rintel <lubo.rintel@gooddata.com>
BUG: 3776