cluster/distribute: readdir - support 64bit offset
ext4 now returns a 64bit offset. In case 64bit offsets are recieved,
our transformation can overflow, and hence lose the offsets.

The new transformation generates the offsets based on shifting left the
offset 2^<num-bits-required for subvolume>. The overflow bits are saved
in fd_ctx for fuse, and passed along in xdata to NFS.

Signed-off-by: shishir gowda <sgowda@redhat.com>
Change-Id: Ia8bb45ad5dc3097277107bb497d3930ec24d7170
BUG: 838784